解决撤销重做bug； 解决yarray删除bug；

+++ lib/utils/struct_store.dart
@@ -247,7 +247,7 @@ class StructStore {
       }
 
       item = find(nextId!);
-      diff = (nextId.clock - item.id.clock).toInt();
+      diff = (nextId.clock - item.id.clock);
       nextId = (item as Item?)?.redone;
     } while (nextId != null && item is Item);
